# decode uuid with builtin UUID type