more clear, but use more memory.