@RX14 2018-01-22
It doesn't use more memory - `UUID` is a struct.