@pynixwang 2018-01-22
@RX14 but allocate UInt8[16]