@RX14 2018-01-22
That's also on the stack. Llvm will likely inline the constructor and to_s and elide the memory copy and you get the same code.