@RX14 2018-01-23
However, returning UUIDs as the new `UUID` struct would be something to discuss... Not sure it's worth it.