@pynixwang 2018-01-23
@RX14 UUID always use as `String`