@pynixwang 2018-01-23
some property such as `version` may never be used